Struct parg::Arg

pub struct Arg { /* fields omitted */ }

This structure represents an Argument for the command line in the form "--arg_name value".

Implementations

impl Arg

pub fn get_name(&self) -> String

Get the name of the Arg.

Example

// match the optional i32 argument --foo <value>
let arg = Arg::with_value("foo", Type::ReadAsI32, false);
assert_eq!(arg.get_name(), String::from("foo"));

pub fn set_description(&self, description: &str)

Sets the Arg description.

Arguments

description - The description of the argument.

Example

let a = Arg::with_value("config", Type::ReadAsString, true);
a.set_description("The argument description");

pub fn with_default_value(
name: &str,
reading_type: Type,
default_value: Box<dyn Any>,
required: bool
) -> Arg

Construct an Arg expecting a value and having a default one.

Arguments

name - The name of the argument.

reading_type - The expected Type of the argument.

default_value - The default value of the argument.

required - Check if whether or not the argument is required.

Example

// match the optional i32 argument --foo <value>
let arg = Arg::with_default_value("foo", Type::ReadAsI32, Box::new(42i32), false);

pub fn with_value(name: &str, reading_type: Type, required: bool) -> Arg

Construct an Arg expecting a value.

Arguments

name - The name of the argument.

reading_type - The expected Type of the argument.

required - Check if whether or not the argument is required.

Example

// match the optional i32 argument --foo <value>
let arg = Arg::with_value("foo", Type::ReadAsI32, false);

pub fn without_value(name: &str, required: bool) -> Arg

Construct an Arg expecting no value.

Arguments

name - The name of the argument.

required - Check if whether or not the argument is required.

Example

// match the optional argument --foo
let arg = Arg::without_value("foo", false);